requirement.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.1
Django==4.1.7
django-adminlte-3==0.1.6
django-cors-headers==3.13.0
django-filter==22.1
django-iranian-cities==1.0.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Markdown==3.4.1
mysqlclient==2.1.1
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
